Split Serialiable/Arrayizable into two classes
